doc: document wheel detection behaviour and move it to be generalized later
